+++ editor/menu/component_viewer.cpp
@@ -12,27 +12,27 @@ void Menu::initComponentViewers() {
       ImGui::PushItemWidth(75);
       ImGui::Text("Position");
       ImGui::SameLine(100);
-      if (ImGui::InputDouble("##transform.position.x", &transform->position.x())) {}
+      if (ImGui::InputFloat("##transform.position.x", &transform->position.x())) {}
       ImGui::SameLine(180);
-      if (ImGui::InputDouble("##transform.position.y", &transform->position.y())) {}
+      if (ImGui::InputFloat("##transform.position.y", &transform->position.y())) {}
       ImGui::SameLine(260);
-      if (ImGui::InputDouble("##transform.position.z", &transform->position.z())) {}
+      if (ImGui::InputFloat("##transform.position.z", &transform->position.z())) {}
 
       ImGui::Text("Rotation");
       ImGui::SameLine(100);
-      if (ImGui::InputDouble("##transform.rotation.x", &transform->rotation.x())) {}
+      if (ImGui::InputFloat("##transform.rotation.x", &transform->rotation.x())) {}
       ImGui::SameLine(180);
-      if (ImGui::InputDouble("##transform.rotation.y", &transform->rotation.y())) {}
+      if (ImGui::InputFloat("##transform.rotation.y", &transform->rotation.y())) {}
       ImGui::SameLine(260);
-      if (ImGui::InputDouble("##transform.rotation.z", &transform->rotation.z())) {}
+      if (ImGui::InputFloat("##transform.rotation.z", &transform->rotation.z())) {}
 
       ImGui::Text("Scale");
       ImGui::SameLine(100);
-      if (ImGui::InputDouble("##transform.scale.x", &transform->scale.x())) {}
+      if (ImGui::InputFloat("##transform.scale.x", &transform->scale.x())) {}
       ImGui::SameLine(180);
-      if (ImGui::InputDouble("##transform.scale.y", &transform->scale.y())) {}
+      if (ImGui::InputFloat("##transform.scale.y", &transform->scale.y())) {}
       ImGui::SameLine(260);
-      if (ImGui::InputDouble("##transform.scale.z", &transform->scale.z())) {}
+      if (ImGui::InputFloat("##transform.scale.z", &transform->scale.z())) {}
     }
   };
 
@@ -41,13 +41,13 @@ void Menu::initComponentViewers() {
     if (ImGui::CollapsingHeader("Physic")) {
       ImGui::Text("Mass");
       ImGui::SameLine(100);
-      ImGui::InputDouble("##physic_mass", &physic->mass);
+      ImGui::InputFloat("##physic_mass", &physic->mass);
       ImGui::Text("Drag");
       ImGui::SameLine(100);
-      ImGui::InputDouble("##physic_drag", &physic->drag);
+      ImGui::InputFloat("##physic_drag", &physic->drag);
       ImGui::Text("Angular drag");
       ImGui::SameLine(100);
-      ImGui::InputDouble("##physic_angular_drag", &physic->angularDrag);
+      ImGui::InputFloat("##physic_angular_drag", &physic->angularDrag);
       ImGui::Text("Gravity");
       ImGui::SameLine(100);
       ImGui::Checkbox("##physic_gravity", &physic->gravity);
